requirements.txt
python-dotenv
mysql-connector-python
supabase
psycopg2-binary
paramiko
